+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItemTooltips.java
@@ -5,6 +5,8 @@ import com.CuteNekoDragon.Core.utils.TooltipBuilder;
 public class SVOItemTooltips {
 
     public static void ProvideTooltips() {
+
+        //Foraged Minerals
         TooltipBuilder.addTooltip(SVOItems.QUARTZ).addCoins().addGlyph(TooltipBuilder.SVOTypes.MINERAL)
                 .addInfo("tooltip.svo_core.quartz");
         TooltipBuilder.addTooltip(SVOItems.EARTH_CRYSTAL).addCoins().addGlyph(TooltipBuilder.SVOTypes.MINERAL)
@@ -13,5 +15,9 @@ public class SVOItemTooltips {
                 .addInfo("tooltip.svo_core.frozen_tear");
         TooltipBuilder.addTooltip(SVOItems.FIRE_QUARTZ).addCoins().addGlyph(TooltipBuilder.SVOTypes.MINERAL)
                 .addInfo("tooltip.svo_core.fire_quartz");
+
+        //Gemstones
+        TooltipBuilder.addTooltip(SVOItems.EMERALD).addCoins().addGlyph(TooltipBuilder.SVOTypes.GEMSTONE)
+                .addInfo("tooltip.svo_core.emerald");
     }
 }