+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItems.java
@@ -52,15 +52,17 @@ public class SVOItems {
         }
     }
 
+    //Foraged Minerals
     public static ItemEntry<Item> QUARTZ = REGISTRATE.item("quartz", Item::new).lang("Quartz")
             .defaultModel().register();
-
     public static ItemEntry<Item> EARTH_CRYSTAL = REGISTRATE.item("earth_crystal", Item::new).lang("Earth Crystal")
             .defaultModel().register();
-
     public static ItemEntry<Item> FROZEN_TEAR = REGISTRATE.item("frozen_tear", Item::new).lang("Frozen Tear")
             .defaultModel().register();
-
     public static ItemEntry<Item> FIRE_QUARTZ = REGISTRATE.item("fire_quartz", Item::new).lang("Fire Quartz")
+            .defaultModel().register();
+
+    //Gemstones
+    public static ItemEntry<Item> EMERALD = REGISTRATE.item("emerald", Item::new).lang("Emerald")
             .defaultModel().register();
 }