+++ changelog.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
 ### Changes
 - changed tooltip color for item types to gold/orange instead of yellow (thanks Randomouse)
+- added gemstone tooltip glyph
+- refactored priceutil to include price types (will be used for price modifiers)
+- added emerald
 
 ## [0.0.16]
 ### Changes